requirements.txt
Flask==3.1.0
mysql_connector_repackaged==0.3.1
mysql-connector-python==8.4.0
PyYAML==6.0.2
waitress==3.0.2
tqdm==4.67.1
nltk==3.9.1